condition on profile to delete indices

+++ clea-integration-tests/src/main/java/fr/gouv/clea/integrationtests/cucumber/CucumberHooks.java
@@ -5,17 +5,22 @@ import lombok.RequiredArgsConstructor;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
+import org.springframework.core.env.Environment;
+import org.springframework.core.env.Profiles;
 
 import java.io.IOException;
 
-// TODO: rework tests so that they don't work only with empty ES indices
 @RequiredArgsConstructor
 public class CucumberHooks {
 
     private final RestHighLevelClient esClient;
 
+    private final Environment environment;
+
     @Before
     public void clearElasticIndices() throws IOException {
-        esClient.indices().delete(new DeleteIndexRequest("*"), RequestOptions.DEFAULT);
+        if (!environment.acceptsProfiles(Profiles.of("int"))) {
+            esClient.indices().delete(new DeleteIndexRequest("*"), RequestOptions.DEFAULT);
+        }
     }
 }
